+++ src/lib.rs
@@ -45,12 +45,11 @@ pub mod sink;
 /// Type alias for a task stored in mysql backend
 pub type MySqlTask<Args> = Task<Args, MySqlContext, Ulid>;
 pub use apalis_sql::config::Config;
+pub use apalis_sql::ext::TaskBuilderExt;
 pub use shared::{SharedMySqlError, SharedMySqlStorage};
 
 pub type MySqlTaskId = apalis_core::task::task_id::TaskId<Ulid>;
 pub type MySqlContext = SqlContext<MySqlPool>;
-
-pub use apalis_sql::ext::TaskBuilderExt;
 
 /// CompactType is the type used for compact serialization in mysql backend
 pub type CompactType = Vec<u8>;
@@ -276,7 +275,7 @@ where
     type CompactStream = TaskStream<MySqlTask<Self::Compact>, sqlx::Error>;
 
     fn get_queue(&self) -> Queue {
-        self.config.queue().to_owned()
+        self.config.queue().clone()
     }
 
     fn poll_compact(self, worker: &WorkerContext) -> Self::CompactStream {
@@ -311,7 +310,6 @@ mod tests {
 
     use apalis::prelude::*;
     use apalis_workflow::*;
-    use chrono::Local;
     use serde::{Deserialize, Serialize};
     use sqlx::MySqlPool;
 
@@ -335,8 +333,6 @@ mod tests {
         })
         .take(ITEMS);
         backend.push_stream(&mut items).await.unwrap();
-
-        println!("Starting worker at {}", Local::now());
 
         async fn send_reminder(item: usize, wrk: WorkerContext) -> Result<(), BoxDynError> {
             if ITEMS == item {